1.8.8 Test

+++ DirectQ/input.h
@@ -25,8 +25,8 @@ void IN_Shutdown (void);
 void IN_ReadRawInput (HRAWINPUT ri_Handle);
 
 void IN_Commands (void);
-void IN_JoyMove (usercmd_t *cmd, float movetime);
-void IN_MouseMove (usercmd_t *cmd, float movetime);
+void IN_JoyMove (usercmd_t *cmd, double movetime);
+void IN_MouseMove (usercmd_t *cmd, double movetime);
 
 void IN_ClearMouseState (void);
 void IN_SetMouseState (bool fullscreen);
